seraphis_crypto: split BulletproofPlus2 into proof-only and complete structs
In `SpTxSquashedV1`, we have commitments (sp enote image and output), and we
have a balance proof. However, the commitments were stored doubly, inside the
in/out lists, and inside the proof. This PR splits up the Bulletproof into
two parts: the values we are committing to/verifying, and the proof itself.
Now inside the `SpBalanceProofV1`, we store just the proof, and not the
committed values. Doing this has many positive effects for code quality.
First, it is now immediately clear to a new reader that the balance proof
won't hold commitments; they "belong" to either an enote image or an output.
Secondly, downstream developers cannot set bulletproof commitments points,
so they can't screw up the 1/8 multiply. Thirdly, this setup makes the
transaction format more "rigid": there are fewer fields that conflict with
each other and there is less room for making silly mistakes based on
assumptions that all fields are important and will change outputs of
functions. For a real world example of what I mean, look at PR fix
https://github.com/monero-project/monero/pull/8707. There was a cache coded
for RingCT signatures with the assumption that changing certain fields of
the RingCT signatures would lead to a different serialization, and thus
changing fields would invalidate the cache. This was NOT the case, and the
cache would have added in an inflation bug if not for a technicality in
the way that RingCT signatures were deserialized and "expanded" using
different elements of the `cryptonote::transaction` class. If
`cryptonote::transaction` perhaps had a more rigid way of constructing
the signatures, it wouldn't have been a problem in the first place. This
is the main goal of this PR: minimize potential future risky behavior.