SpTxSquashedV1: simplify txid
We got rid of the tx artifacts hash. The reason for it was a follows:
```
note: key images are represented in the tx id twice (tx proposal prefix and input images
-    //   - the reasons are: A) decouple proposals from the enote image structure, B) don't require proposals to commit
-    //     to input commitment masks
```

Input proposals DO commit to input commitment masks though (see `SpInputProposalCore`). Does the txid hash indirection
still serve a purpose or could we make the txid faster?